EMR Cashier module

+++ api/src/main/java/org/openmrs/module/kenyaemr/cashier/api/base/PagingInfo.java
@@ -73,6 +73,6 @@ public class PagingInfo {
 	}
 
 	public Boolean hasMoreResults() {
-		return (page * pageSize) < totalRecordCount;
+		return ((long)page * pageSize) < totalRecordCount;
 	}
 }
